+++ selftests/unit/test_multiplexer.py
@@ -4,6 +4,7 @@ import sys
 
 from avocado.core import multiplexer
 from avocado.core import tree
+from avocado.plugins import yaml_to_mux
 
 if sys.version_info[:2] == (2, 6):
     import unittest2 as unittest
@@ -29,23 +30,19 @@ class TestMultiplex(unittest.TestCase):
                                                'examples/mux-selftest.yaml'])
         self.mux_full = tuple(multiplexer.MuxTree(self.mux_tree))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_empty(self):
         act = tuple(multiplexer.MuxTree(tree.TreeNode()))
         self.assertEqual(act, (['', ],))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_partial(self):
         exp = (['intel', 'scsi'], ['intel', 'virtio'], ['amd', 'scsi'],
                ['amd', 'virtio'], ['arm', 'scsi'], ['arm', 'virtio'])
         act = tuple(multiplexer.MuxTree(self.mux_tree.children[0]))
         self.assertEqual(act, exp)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_full(self):
         self.assertEqual(len(self.mux_full), 12)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_create_variants(self):
         from_file = multiplexer.yaml2tree(
             ["/:" + PATH_PREFIX + 'examples/mux-selftest.yaml'])
@@ -53,7 +50,7 @@ class TestMultiplex(unittest.TestCase):
         self.assertEqual(self.mux_full, tuple(from_file))
 
     # Filters are tested in tree_unittests, only verify `multiplex_yamls` calls
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_filter_only(self):
         exp = (['intel', 'scsi'], ['intel', 'virtio'])
         act = multiplexer.yaml2tree(["/:" + PATH_PREFIX +
@@ -64,7 +61,7 @@ class TestMultiplex(unittest.TestCase):
         act = tuple(multiplexer.MuxTree(act))
         self.assertEqual(act, exp)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_filter_out(self):
         act = multiplexer.yaml2tree(["/:" + PATH_PREFIX +
                                      'examples/mux-selftest.yaml'],
@@ -95,13 +92,13 @@ class TestAvocadoParams(unittest.TestCase):
         self.params2 = multiplexer.AvocadoParams(self.yamls.next(), 'Unittest2',
                                                  ['/ch1/*', '/ch0/*'], {})
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_pickle(self):
         params = pickle.dumps(self.params1, 2)  # protocol == 2
         params = pickle.loads(params)
         self.assertEqual(self.params1, params)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_basic(self):
         self.assertEqual(self.params1, self.params1)
         self.assertNotEqual(self.params1, self.params2)
@@ -110,7 +107,7 @@ class TestAvocadoParams(unittest.TestCase):
         str(multiplexer.AvocadoParams([], 'Unittest', [], {}))
         self.assertEqual(15, sum([1 for _ in self.params1.iteritems()]))
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_unhashable(self):
         """ Verifies that unhashable arguments can be passed to params.get """
         self.assertEqual(self.params1.get("root", "/ch0/", ["foo"]), ["foo"])
@@ -118,7 +115,7 @@ class TestAvocadoParams(unittest.TestCase):
                                           '/ch0/ch0.1/ch0.1.1/ch0.1.1.1/',
                                           ['bar']), 'unique1')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_abs_path(self):
         # /ch0/ is not leaf thus it's not queryable
         self.assertEqual(self.params1.get('root', '/ch0/', 'bbb'), 'bbb')
@@ -140,7 +137,7 @@ class TestAvocadoParams(unittest.TestCase):
                                           '/ch0/ch0.1/ch0.1.1/ch0.1.1.1/',
                                           'hhh'), 'hhh')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_greedy_path(self):
         self.assertEqual(self.params1.get('unique1', '/*/*/*/ch0.1.1.1/',
                                           111), 'unique1')
@@ -164,7 +161,7 @@ class TestAvocadoParams(unittest.TestCase):
         # path matches nothing
         self.assertEqual(self.params1.get('root', '', 999), 999)
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_rel_path(self):
         self.assertEqual(self.params1.get('root', default='iii'), 'root')
         self.assertEqual(self.params1.get('unique1', '*', 'jjj'), 'unique1')
@@ -179,7 +176,7 @@ class TestAvocadoParams(unittest.TestCase):
         self.assertEqual(self.params2.get('unique1', '*/ch0.1.1.1/', 'ooo'),
                          'ooo')
 
-    @unittest.skipIf(not tree.MULTIPLEX_CAPABLE, "Not multiplex capable")
+    @unittest.skipIf(not yaml_to_mux.MULTIPLEX_CAPABLE, "Not multiplex capable")
     def test_get_clashes(self):
         # One inherited, the other is new
         self.assertRaisesRegexp(ValueError, r"'clash1'.* \['/ch0/ch0.1/ch0.1.1"
